requirements.txt
paho-mqtt==1.6.1
PySimpleGUI==4.60.5
